Code full dto, respository

+++ src/main/java/com/FTimeshare/UsageManagement/repositories/NewsRepository.java
@@ -1,8 +1,7 @@
 package com.FTimeshare.UsageManagement.repositories;
 
-
 import com.FTimeshare.UsageManagement.entities.NewsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NewsRepository extends JpaRepository<NewsEntity, Integer> {
+public interface NewsRepository extends JpaRepository <NewsEntity, Long> {
 }
